+++ RELEASES.md
@@ -1,11 +1,18 @@
 # Release Notes
 
-## Version 0.3.1
+## Version 0.4
 
 ### Usability
 
 - reduced required `derive_more` features
 - removed `thiserror` dependency
+
+### Bug fixes
+
+- fixed serious bug that broke all functionality relating to durations that buttons were pressed or released for
+  - `ActionState::tick` now takes the `Instant` of both the current and previous frame, rather than just the current
+- `InputManagerPlugin` no longer panics when time does not have a previous update
+  - this is useful as it ensures `bevy_inspector_egui` compatibility!
 
 ### Docs